System-1 split subtests to individual tests (#3804)
* Create system_banner_test.go

* Create README.md - banner_test.go

* Create metadata.textproto-banner test

* Create system_g_protocol_test.go

* Create README.md-g protocol test

* Create metadata.textproto

* Update README.md

* Update README.md

* Create hostname_test.go

* Create README.md

* Create metadata.textproto

* Create system_time_test.go

* Create README.md

* Update README.md

* Create metadata.textproto

* Update hostname_test.go

* Update system_banner_test.go

* Update system_g_protocol_test.go

* Update system_time_test.go

* Delete feature/system/tests/system_base_test/time_test.go

* Delete feature/system/tests/system_base_test/system_base_test.go

* Delete feature/system/tests/system_base_test/metadata.textproto

* Delete feature/system/tests/system_base_test/hostname_test.go

* Delete feature/system/tests/system_base_test/g_protocol_test.go

* Delete feature/system/tests/system_base_test/banner_test.go

* Delete feature/system/tests/system_base_test/README.md

* Update README.md

* Update metadata.textproto

* Update metadata.textproto

* Update metadata.textproto

* Rename README.md to README.md

* Rename hostname_test.go to hostname_test.go

* Rename metadata.textproto to metadata.textproto

* Rename README.md to README.md

* Rename metadata.textproto to metadata.textproto

* Rename system_banner_test.go to system_banner_test.go

* Rename README.md to README.md

* Rename metadata.textproto to metadata.textproto

* Rename system_g_protocol_test.go to system_g_protocol_test.go

* Rename README.md to README.md

* Rename metadata.textproto to metadata.textproto

* Rename system_time_test.go to system_time_test.go

* Rename feature/system/feature.textproto to feature/system/system_base_test/feature.textproto

* Rename feature/system/system_base_test/feature.textproto to feature/system/feature.textproto

* format hostname.go

* format system_banner_test.go

* format system_banner_test.go

* format hostname.go

* format g_protocol.go

* go formatted

---------
