+++ feature/system/system_base_test/tests/hostname_test/metadata.textproto
@@ -1,9 +1,9 @@
 # proto-file: github.com/openconfig/featureprofiles/proto/metadata.proto
 # proto-message: Metadata
 
-uuid:  "ec59d852-0828-4607-bea5-fbe2340c5414"
-plan_id:  "System-1"
-description:  "System testing"
+uuid:  "4e642f03-c35e-4780-a885-2e44563eef17"
+plan_id:  "System-1.3"
+description:  "System hostname test"
 testbed:  TESTBED_DUT_ATE_2LINKS
 platform_exceptions:  {
   platform:  {
@@ -11,14 +11,5 @@ platform_exceptions:  {
   }
   deviations:  {
     cli_takes_precedence_over_oc:  true
-  }
-}
-platform_exceptions: {
-  platform: {
-    vendor: CISCO
-  }
-  deviations: {
-    p4rt_capabilities_unsupported: true
-    gnmi_get_on_root_unsupported: true
   }
 }
