+++ feature/system/system_base_test/tests/system_banner_test/metadata.textproto
@@ -1,0 +1,7 @@
+# proto-file: github.com/openconfig/featureprofiles/proto/metadata.proto
+# proto-message: Metadata
+
+uuid:  "ec59d852-0828-4607-bea5-fbe2340c5414"
+plan_id:  "System-1.1"
+description:  "System banner test"
+testbed:  TESTBED_DUT_ATE_2LINKS
